requirements.txt
absl-py==2.2.2
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antio==0.0.0
anyio==4.10.0
argon2-cffi==21.3.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
axial_positional_embedding==0.3.12
babel==2.16.0
beautifulsoup4==4.13.5
bleach==6.2.0
Bottleneck==1.4.2
braindecode==0.8.1
Brotli==1.0.9
brotlicffi==1.0.9.2
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
CoLT5-attention==0.11.1
comm==0.2.2
contourpy==1.3.1
cupy-cuda12x==13.4.1
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.11
decorator==5.2.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.18
dgl==2.1.0
dipy==1.11.0
docstring-inheritance==2.2.2
edfio==0.4.9
eeglabio==0.1.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
fastrlock==0.8.3
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
graphgallery==1.1.0
grpcio==1.71.0
h11==0.16.0
h5io==0.2.5
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
hyper-connections==0.2.1
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.9.25
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.9.2
keras_sequential_ascii==0.1.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
LibEER==0.1.0+ako
linear-attention-transformer==0.19.1
linformer==0.2.3
llvmlite==0.45.0
local-attention==1.11.2
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mat73==0.65
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mffpy==0.10.0
mistune==3.1.2
ml_dtypes==0.5.1
mne==1.10.1
mne-bids==0.17.0
mne-qt-browser==0.7.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.12.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.62.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
opencv-python==4.12.0.88
openi==2.0.4
openmeeg==2.5.12
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
ordered-set==4.1.0
orjson==3.10.14
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.0
platformdirs==4.3.7
ply==3.11
pooch==1.8.2
product_key_memory==0.2.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.0
ptflops==0.7.5
ptyprocess==0.7.0
pure_eval==0.2.3
pwinput==1.0.3
py-cpuinfo==9.0.0
pybv==0.7.6
pycares==4.10.0
pycparser==2.23
Pygments==2.19.1
pymatreader==1.1.0
PyOpenGL==3.1.9
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyqtgraph==0.13.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-picard==0.8
pytorch-triton==3.3.0+git96316ce5
pytorch-warmup==0.2.0
pytz==2024.1
pyvista==0.46.3
pyvistaqt==0.11.3
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qtconsole==5.7.0
QtPy==2.4.3
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.15.2
scooby==0.10.2
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==75.8.0
setuptools-scm==8.3.1
sip==6.7.12
six==1.17.0
skorch==1.2.0
smart_open==7.3.1
sniffio==1.3.0
soundfile==0.13.1
soupsieve==2.5
spikingjelly==0.0.0.0.14
stack_data==0.6.3
statsmodels==0.14.5
sympy==1.13.3
tables==3.10.2
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.0.1
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
torch==2.8.0.dev20250330+cu128
torch-geometric==2.6.1
torchaudio==2.6.0.dev20250330+cu128
torchdata==0.11.0
torchinfo==1.8.0
torchvision==0.22.0.dev20250330+cu128
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trame==3.12.0
trame-client==3.11.0
trame-common==1.0.1
trame-server==3.6.2
trame-vtk==2.10.0
trame-vuetify==3.1.0
trx-python==0.3
typing_extensions==4.13.2
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.3.0
uvloop==0.21.0
vtk==9.3.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wfdb==4.3.0
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wslink==2.4.0
xlrd==2.0.1
xmltodict==1.0.2
yacs==0.1.8
yarl==1.18.0
zipp==3.21.0
